% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{calculate_grades_with_exceptions}
\alias{calculate_grades_with_exceptions}
\title{Calculate Grades with Exceptions}
\usage{
calculate_grades_with_exceptions(policies)
}
\arguments{
\item{policies}{A list with separate policies and relevant gs}
}
\value{
A list with various policies and truncated gs
}
\description{
Apply each policy to respective students and combine all grades
}
