% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{get_grades_with_exceptions}
\alias{get_grades_with_exceptions}
\title{Get Grades With Exceptions}
\usage{
get_grades_with_exceptions(gs = gs, policy = policy)
}
\arguments{
\item{gs}{A Gradescope dataframe}

\item{policy}{A grading policy file as an R list}
}
\value{
A dataframe of the original Gradescope data with computed categories' scores appended as additional columns
}
\description{
Get grades with exceptions applied to relevant students.
}
