% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_canvas}
\alias{read_canvas}
\title{This function takes in gradebook csv and lateness csv from Canvas
and aggregates them into identical format as those from Gradescope.
Note that Canvas defaults assignment due time to beginning of day in lateness csv
regardless of true time of lateness.}
\usage{
read_canvas(path_to_grades, path_to_lateness, verbose = FALSE)
}
\arguments{
\item{path_to_grades}{The path to Canvas gradebook csv}

\item{path_to_lateness}{The path to Canvas csv for assignment Lateness}
}
\value{
A tibble in correct format a la Gradescope
}
\description{
Data is then checked for correct format.
}
