% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{separate_policies}
\alias{separate_policies}
\title{Separate policies}
\usage{
separate_policies(policy, gs)
}
\arguments{
\item{policy}{A grading policy file as an R list}

\item{gs}{A Gradescope dataframe}
}
\value{
A list with various policies and truncated gs
}
\description{
Create separate policies for each unique exception and truncate gs based on
which students these exceptions apply to
}
